Rebrand to Dayli
The concept of using the Zulu/Xhosa word for a journal,
"Ijenali", was appealing to bring an African feel to an essentially
African challenge.

Visually it didn't work very well as an app name.

Dayli works nicely as it is a short word, a play on the daily
activity of journaling, and is more unique for SEO purposes.

This commit also includes some copy updates, and
screenshots of the new branding.